Make Control writing consistant
gob encoding of maps does not result in consistant data.
We store as seperate arrays of keys and values, so that
hashes of the data are consistant.

We also make sure that we write the unordered fields of
a control file out in alphabetic order when displaying,
or generating the Packages and Sources files